requirements.txt
appdirs==1.4.4
asgiref==3.3.4
black==21.5b0
click==7.1.2
coverage==5.5
Django==3.2.2
django-filter==2.4.0
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.1.2
flake8==3.9.1
flake8-variables-names==0.0.4
isort==5.8.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
